# server/openapi.yaml
openapi: 3.0.0
info:
  title: Blog Poetry API
  version: 1.0.0
  description: API for managing Poems, Galleries, and User interactions.
servers:
  - url: http://localhost:3001
    description: Local server
paths:
...
  /api/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [poet, cititor]
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input
        '409':
          description: Email already exists
  /api/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [poet, cititor]
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
...